fixed instruction set encoding (all items are aligned for 9 bytes), fixed poseidon hash calculation, fixed incorrect halo2 version

+++ rwasm/src/engine/bytecode/mod.rs
@@ -23,7 +23,7 @@ pub use self::utils::{
 use super::{const_pool::ConstRef, CompiledFunc, TranslationError};
 use crate::common::{UntypedValue, F32};
 use core::fmt::Debug;
-#[cfg(test)]
+#[cfg(feature = "std")]
 use strum_macros::EnumIter;
 
 /// The internal `wasmi` bytecode that is stored for Wasm functions.
@@ -35,7 +35,7 @@ use strum_macros::EnumIter;
 /// For example the `BrTable` instruction is unrolled into separate instructions
 /// each representing either the `BrTable` head or one of its branching targets.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-#[cfg_attr(test, derive(EnumIter))]
+#[cfg_attr(feature = "std", derive(EnumIter))]
 pub enum Instruction {
     LocalGet(LocalDepth),
     LocalSet(LocalDepth),
@@ -348,7 +348,7 @@ pub enum Instruction {
     I64TruncSatF64S,
     I64TruncSatF64U,
     // sanitizer params
-    SanitizerStackCheck(i32),
+    SanitizerStackCheck(i64),
 }
 
 impl Instruction {
